* samtools-0.1.4-16 (r360)  * report more information in index when the input is not sorted  * change the behaviour of knet_seek() such that it returns 0 on success  * support knetfile library in BGZF

+++ knetfile.h
@@ -30,7 +30,7 @@ extern "C" {
 	knetFile *knet_open(const char *fn, const char *mode);
 	knetFile *knet_dopen(int fd, const char *mode);
 	off_t knet_read(knetFile *fp, void *buf, off_t len);
-	off_t knet_seek(knetFile *fp, off_t off, int whence);
+	int knet_seek(knetFile *fp, off_t off, int whence);
 	int knet_close(knetFile *fp);
 
 #ifdef __cplusplus
